refactor(reservations): split views into different files
