+++ apps/reservations/models/__init__.py
@@ -1,6 +1,6 @@
+from .area import Area
+from .rating import Rating
+from .reservation import Reservation
 from .restaurant import Restaurant
 from .table import Table, TableSchedule
-from .area import Area
 from .turn import Turn
-from .reservation import Reservation
-from .rating import Rating
