+++ apps/reservations/serializers/__init__.py
@@ -1,6 +1,6 @@
+from .area import AreaSerializer
+from .rating import RatingSerializer
+from .reservation import ReservationsSerializer
 from .restaurant import RestaurantSerializer
 from .table import TableSerializer, TableScheduleSerializer
-from .area import AreaSerializer
 from .turn import TurnSerializer
-from .reservation import ReservationsSerializer
-from .rating import RatingSerializer
