+++ apps/reservations/views/__init__.py
@@ -1,0 +1,6 @@
+from .area import AreaViewSet
+from .rating import RatingViewSet
+from .reservation import ReservationViewSet
+from .restaurant import RestaurantViewSet
+from .table import TableViewSet, TableScheduleViewSet
+from .turn import TurnViewSet
