+++ apps/reservations/views/area.py
@@ -1,0 +1,11 @@
+from rest_framework import viewsets
+from rest_framework.permissions import IsAdminUser, IsAuthenticated
+from drf_yasg.utils import swagger_auto_schema
+from apps.reservations.models import Area
+from apps.reservations.serializers import AreaSerializer
+
+
+class AreaViewSet(viewsets.ModelViewSet):
+    queryset = Area.objects.all()
+    serializer_class = AreaSerializer
+    permission_classes = [IsAdminUser]
