+++ apps/reservations/views/rating.py
@@ -1,0 +1,11 @@
+from rest_framework import viewsets
+from rest_framework.permissions import IsAdminUser, IsAuthenticated
+from drf_yasg.utils import swagger_auto_schema
+from apps.reservations.models import Rating
+from apps.reservations.serializers import RatingSerializer
+
+
+class RatingViewSet(viewsets.ModelViewSet):
+    queryset = Rating.objects.all()
+    serializer_class = RatingSerializer
+    permission_classes = [IsAdminUser, IsAuthenticated]
