+++ apps/reservations/views/restaurant.py
@@ -1,0 +1,11 @@
+from rest_framework import viewsets
+from rest_framework.permissions import IsAdminUser, IsAuthenticated
+from drf_yasg.utils import swagger_auto_schema
+from apps.reservations.models import Restaurant
+from apps.reservations.serializers import RestaurantSerializer
+
+
+class RestaurantViewSet(viewsets.ModelViewSet):
+    queryset = Restaurant.objects.all()
+    serializer_class = RestaurantSerializer
+    permission_classes = [IsAdminUser]
